create seeders ingredients+pizzas and migrate to database

+++ database/migrations/2023_03_30_111630_create_ingredients_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('ingredients', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('cost_price');
+            $table->decimal('cost_price',8,2);
             $table->timestamps();
         });
     }
